+++ samples/SampleApi/ServiceAgents/OAuthDemoAgent.cs
@@ -1,19 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Options;
-using Digipolis.ServiceAgents;
+﻿using Digipolis.ServiceAgents;
 using Digipolis.ServiceAgents.Settings;
-using SampleApi.Models;
-using System.Threading;
+using Microsoft.Extensions.Options;
+using System;
+using System.Net.Http;
+using System.Threading.Tasks;
 
 namespace SampleApi.ServiceAgents
 {
     public class OAuthDemoAgent : AgentBase, IOAuthDemoAgent
     {
-        public OAuthDemoAgent(IServiceProvider serviceProvider, IOptions<ServiceAgentSettings> options) 
-            : base(serviceProvider, options)
+        public OAuthDemoAgent(HttpClient client, IServiceProvider serviceProvider, IOptions<ServiceAgentSettings> options) 
+            : base(client, serviceProvider, options)
         {
         }