+++ test/Digipolis.ServiceAgents.UnitTests/Startup/AddSingleServiceAgentTests.cs
@@ -3,6 +3,7 @@ using Digipolis.ServiceAgents.Settings;
 using Digipolis.ServiceAgents.UnitTests.Utilities;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Moq;
 using System;
 using System.Linq;
 using System.Net.Http;
@@ -21,36 +22,27 @@ namespace Digipolis.ServiceAgents.UnitTests.Startup
 
             var ex = Assert.Throws<ArgumentNullException>(() => services.AddSingleServiceAgent<TestAgent>(nullAction));
 
-            Assert.Equal("setupAction", ex.ParamName);
+            Assert.Equal("serviceSettingsSetupAction", ex.ParamName);
         }
 
         [Fact]
-        private void HttpClientFactoryIsRegistratedAsSingleton()
+        private void HttpClientCreatedActionIsExecuted()
         {
-            var services = new ServiceCollection();
-            services.AddSingleServiceAgent<TestAgent>(settings => { });
-
-            var registrations = services.Where(sd => sd.ServiceType == typeof(IHttpClientFactory))
-                                        .ToArray();
-
-            Assert.Equal(1, registrations.Count());
-            Assert.Equal(ServiceLifetime.Singleton, registrations[0].Lifetime);
-        }
-
-        [Fact]
-        private void HttpClientFactoryClientActionIsPassed()
-        {
-            var serviceAgentSettings = new ServiceAgentSettings();
             HttpClient passedClient = null;
             var services = new ServiceCollection();
-            services.AddSingleServiceAgent<TestAgent>(settings => { }, (sp, client) => passedClient = client);
+            services.AddSingleServiceAgent<TestAgent>(settings => {
+                settings.Scheme = "http";
+                settings.Host = "localhost";
+                settings.Path = "api";
+            }, 
+            (sp, client) => {
+                // execution of the HttpClientCreated action
+                passedClient = client;
+            });
 
-            ///get the registrated HttpFactory
-            var registration = services.Single(sd => sd.ServiceType == typeof(IHttpClientFactory));
-
-            //Manually call the CreateClient on the factory (this normally happens when the service agent gets resolved
-            var factory = registration.ImplementationFactory.Invoke(null) as HttpClientFactory;
-            factory.CreateClient(serviceAgentSettings, new ServiceSettings { Host = "test.be" });
+            /// get the registrated TestAgent > this also creates an HttpClient and executes the clientCreatedAction
+            var registration = services.Single(sd => sd.ServiceType == typeof(TestAgent));
+            var agent = registration.ImplementationFactory.Invoke(services.BuildServiceProvider()) as TestAgent;
 
             Assert.NotNull(passedClient);
         }
@@ -71,7 +63,7 @@ namespace Digipolis.ServiceAgents.UnitTests.Startup
             var registrations = services.Where(sd => sd.ServiceType == typeof(IConfigureOptions<ServiceAgentSettings>))
                                         .ToArray();
 
-            Assert.Equal(1, registrations.Count());
+            Assert.Single(registrations);
             Assert.Equal(ServiceLifetime.Singleton, registrations[0].Lifetime);
 
             var configOptions = registrations[0].ImplementationInstance as IConfigureOptions<ServiceAgentSettings>;
@@ -93,33 +85,32 @@ namespace Digipolis.ServiceAgents.UnitTests.Startup
         }
 
         [Fact]
-        private void ServiceAgentIsRegistratedAsScoped()
+        private void ServiceAgentIsRegistratedAsTransient()
         {
             var services = new ServiceCollection();
             services.AddSingleServiceAgent<TestAgent>(settings => { });
 
-            var registrations = services.Where(sd => sd.ServiceType == typeof(TestAgent) &&
-                                                     sd.ImplementationType == typeof(TestAgent))
+            var registrations = services.Where(sd => sd.ServiceType == typeof(TestAgent))
                                         .ToArray();
 
-            Assert.Equal(1, registrations.Count());
-            Assert.Equal(ServiceLifetime.Scoped, registrations[0].Lifetime);
+            Assert.Single(registrations);
+            Assert.Equal(ServiceLifetime.Transient, registrations[0].Lifetime);
         }
 
-        [Fact]
-        private void ServiceAgentInterfaceIsRegistratedAsScoped()
-        {
-            var services = new ServiceCollection();
-            services.AddSingleServiceAgent<InterfaceImplementingAgent>(settings => { },
-                assembly: typeof(InterfaceImplementingAgent).GetTypeInfo().Assembly);
+        // not a valid unit test for now: AddSingleServiceAgent<T> -> service is registered as T and not with a possible interface
+        //[Fact]
+        //private void ServiceAgentInterfaceIsRegistratedAsTransient()
+        //{
+        //    var services = new ServiceCollection();
+        //    services.AddSingleServiceAgent<InterfaceImplementingAgent>(servicSettings => { },
+        //        assembly: typeof(InterfaceImplementingAgent).GetTypeInfo().Assembly);
 
-            var registrations = services.Where(sd => sd.ServiceType == typeof(IInterfaceImplementingAgent) &&
-                                                     sd.ImplementationType == typeof(InterfaceImplementingAgent))
-                                        .ToArray();
+        //    var registrations = services.Where(sd => sd.ServiceType == typeof(IInterfaceImplementingAgent))
+        //                                .ToArray();
 
-            Assert.Equal(1, registrations.Count());
-            Assert.Equal(ServiceLifetime.Scoped, registrations[0].Lifetime);
-        }
+        //    Assert.Single(registrations);
+        //    Assert.Equal(ServiceLifetime.Transient, registrations[0].Lifetime);
+        //}
 
         [Fact]
         private void TokenHelperIsRegistratedAsScoped()
@@ -132,8 +123,31 @@ namespace Digipolis.ServiceAgents.UnitTests.Startup
                                                      sd.ImplementationType == typeof(TokenHelper))
                                         .ToArray();
 
-            Assert.Equal(1, registrations.Count());
+            Assert.Single(registrations);
             Assert.Equal(ServiceLifetime.Scoped, registrations[0].Lifetime);
+        }
+
+        private ServiceAgentSettings CreateServiceAgentSettings()
+        {
+            var settings = new ServiceAgentSettings();
+            settings.Services.Add("TestAgent", new ServiceSettings { Scheme = HttpSchema.Http, Host = "localhost" });
+            return settings;
+        }
+
+        private IServiceProvider CreateServiceProvider(ServiceAgentSettings settings)
+        {
+            var serviceProviderMock = new Mock<IServiceProvider>();
+
+            if (settings != null)
+                serviceProviderMock.Setup(p => p.GetService(typeof(IOptions<ServiceAgentSettings>))).Returns(Options.Create(settings));
+
+            var authContextMock = new Mock<IAuthContext>();
+            authContextMock.Setup(c => c.UserToken).Returns("TokenValue");
+
+            serviceProviderMock.Setup(p => p.GetService(typeof(IAuthContext))).Returns(authContextMock.Object);
+            serviceProviderMock.Setup(p => p.GetService(typeof(IRequestHeaderHelper))).Returns(new RequestHeaderHelper(serviceProviderMock.Object));
+
+            return serviceProviderMock.Object;
         }
     }
 }